gl_timer，用于背景工作线程调度的timer。MainFrame的OnTimer()只用于更新系统状态任务，数据接收处理等定时调度任务由gl_timer所关联的ScheduleTask()完成。 并行处理线程的concurrencpp::result返回值无法读取。待查。

+++ src/Type/China market/WebData.h
@@ -42,7 +42,7 @@ public:
 	void SetTime(const time_t tTime) noexcept { m_tTime = tTime; }
 	CString GetStockCode() const noexcept { return m_strStockCode; }
 	void SetStockCode(const CString& strStockCode) noexcept { m_strStockCode = strStockCode; }
-	INT64 GetBufferLength() const noexcept { return m_sDataBuffer.size(); }
+	size_t GetBufferLength() const noexcept { return m_sDataBuffer.size(); }
 
 	string GetDataBuffer() noexcept { return m_sDataBuffer; }
 	long GetCurrentPos() const noexcept { return m_lCurrentPos; }
